Refactor energy event proto (Studio-side)
Bug: 129355112
Test: existing

+++ transport-database/src/main/java/com/android/tools/datastore/energy/PowerProfile.java
@@ -261,10 +261,10 @@ public interface PowerProfile {
   }
 
   final class LocationEvent {
-    public final int myEventId;
+    public final long myEventId;
     public final LocationType myLocationType;
 
-    public LocationEvent(int id, @NotNull LocationType type) {
+    public LocationEvent(long id, @NotNull LocationType type) {
       myEventId = id;
       myLocationType = type;
     }
